requirements.txt
pyserial
pycaw
pystray